requirements.txt
absl-py==2.1.0
numpy==1.26.4
scipy==1.10.1
matplotlib==3.9.2
pandas==2.1.1
xarray==2025.1.2
tqdm==4.66.1
opencv-python==4.9.0.80
connected-components-3d==3.15.0
psutil==5.9.5
geopandas==1.0.1
shapely==2.0.2
rasterio==1.3.9
cartopy==0.23.0
netCDF4
h5netcdf==1.4.1
dask==2024.1.1
